+++ include/Util.h
@@ -9,7 +9,7 @@
 #include "GlmHeaders.h"
 #include <SDL2/SDL.h>
 
-const GLfloat DegreesToRadians = M_PI / 180.0;
+const GLdouble DegreesToRadians = M_PI / 180.0;
 
 namespace puddi
 {
