+++ src/ModelGraph.cc
@@ -55,7 +55,7 @@ namespace puddi
 				if (parallel)
 				{
 					#pragma omp parallel for
-					for (int i = 0; i < children.size(); ++i)
+					for (int i = 0; i < static_cast<int>(children.size()); ++i)
 					{
 						//std::cout << "i = " << i << ", threads = " << omp_get_num_threads() << std::endl;
 						object->PassDownParentModel(children[i]->object);
@@ -77,7 +77,7 @@ namespace puddi
 				if (parallel)
 				{
 					#pragma omp parallel for
-					for (int i = 0; i < children.size(); ++i)
+					for (int i = 0; i < static_cast<int>(children.size()); ++i)
 					{
 						children[i]->Update();
 					}
